refactor: share WP Codebox trace recipe runner
