+++ Automattic/studio/bench/lib/wordpress-helper-discovery.mjs
@@ -5,17 +5,6 @@ import path from 'node:path';
 const require = createRequire(import.meta.url);
 const WORDPRESS_HELPER_CONSUMER_FILENAME = 'wordpress-helper-consumer.js';
 
-const WORDPRESS_LIB_HELPER_KEYS = new Map([
-  ['block-quality.js', 'blockQuality'],
-  ['editor-canvas-probes.js', 'editorCanvasProbes'],
-  ['fidelity-comparison.js', 'fidelityComparison'],
-  ['fixture-setup.js', 'fixtureSetup'],
-  ['materialized-site-quality.js', 'materializedSiteQuality'],
-  ['request-profiler.js', 'requestProfiler'],
-  ['timing-correlator.js', 'timingCorrelator'],
-  ['wordpress-bootstrap-timeline.js', 'bootstrapTimeline'],
-]);
-
 function resolveManifestPath(options = {}) {
   return options.manifestPath || process.env.HOMEBOY_WORDPRESS_HELPER_MANIFEST || '';
 }
@@ -85,17 +74,7 @@ export function wordpressLibHelperPath(fileName, options = {}) {
   }
 
   const explicit = options.override || (options.envVar ? process.env[options.envVar] : '');
-  if (explicit) {
-    return explicit;
-  }
-
-  const { manifest } = localManifestForConsumerDiscovery(options);
-  const helperKey = options.helperKey || WORDPRESS_LIB_HELPER_KEYS.get(fileName);
-  if (helperKey && manifest?.helpers?.[helperKey]) {
-    return manifest.helpers[helperKey];
-  }
-
-  return manifest?.extensionRoot ? path.join(manifest.extensionRoot, 'lib', fileName) : '';
+  return explicit || '';
 }
 
 export function loadWordPressLibHelper(fileName, options = {}) {
