+++ WordPress/gutenberg/bench/pattern-preview-assets.trace.mjs
@@ -1,18 +1,14 @@
-import { execFile } from 'node:child_process';
 import { mkdir, mkdtemp, readFile, rm, writeFile } from 'node:fs/promises';
 import { existsSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
-import { promisify } from 'node:util';
-
-const execFileAsync = promisify( execFile );
+import { runWpCodeboxRecipe } from '../../../shared/wp-codebox/recipe.mjs';
 
 const componentPath = process.env.HOMEBOY_COMPONENT_PATH;
 const componentId = process.env.HOMEBOY_COMPONENT_ID || 'gutenberg';
 const scenarioId = process.env.HOMEBOY_TRACE_SCENARIO || 'pattern-preview-assets';
 const resultsFile = process.env.HOMEBOY_TRACE_RESULTS_FILE;
 const artifactDir = process.env.HOMEBOY_TRACE_ARTIFACT_DIR || path.join( tmpdir(), 'gutenberg-pattern-preview-assets-artifacts' );
-const wpCodeboxBin = process.env.HOMEBOY_WP_CODEBOX_BIN || process.env.HOMEBOY_SETTINGS_WP_CODEBOX_BIN || path.join( process.env.HOME || '', 'Developer/wp-codebox/packages/cli/dist/index.js' );
 const wpVersion = process.env.HOMEBOY_GUTENBERG_PATTERN_ASSETS_WP_VERSION || '7.0';
 const blockCount = Number.parseInt( process.env.HOMEBOY_GUTENBERG_PATTERN_ASSETS_BLOCK_COUNT || '6', 10 );
 const patternCount = Number.parseInt( process.env.HOMEBOY_GUTENBERG_PATTERN_ASSETS_PATTERN_COUNT || '12', 10 );
@@ -54,14 +50,6 @@ function timestampMs() {
 
 function event( source, name, data = {} ) {
 	timeline.push( { t_ms: timestampMs(), source, event: name, data } );
-}
-
-function wpCodeboxCommand() {
-	if ( wpCodeboxBin.endsWith( '.js' ) || wpCodeboxBin.endsWith( '.cjs' ) || wpCodeboxBin.endsWith( '.mjs' ) ) {
-		return { command: 'node', args: [ wpCodeboxBin ] };
-	}
-
-	return { command: wpCodeboxBin, args: [] };
 }
 
 async function readJsonAsync( pathname ) {
@@ -488,12 +476,13 @@ try {
 
 	await writeFile( recipeFile, `${ JSON.stringify( recipe, null, 2 ) }\n` );
 
-	event( 'wp_codebox', 'recipe.start', { recipe_file: recipeFile } );
-	const { command, args } = wpCodeboxCommand();
-	const result = await execFileAsync( command, [ ...args, 'recipe-run', '--recipe', recipeFile, '--artifacts', codeboxArtifacts, '--json' ], {
+	const result = await runWpCodeboxRecipe( {
+		recipeFile,
+		artifactsDir: codeboxArtifacts,
+		outputFile,
+		event,
 		maxBuffer: 1024 * 1024 * 50,
 	} );
-	await writeFile( outputFile, result.stdout );
 
 	const output = JSON.parse( result.stdout );
 	const bundleDir = output.artifacts?.directory;
